requirements.txt
# Core ML and Deep Learning
tensorflow==2.9.1
torch==2.0.0
torchvision==0.15.0

# Image Processing
opencv-python==4.8.0.76
Pillow==9.5.0
albumentations==1.3.0

# Data Processing
numpy==1.24.0
pandas==2.0.0
pycocotools==2.0.6

# Web Interface
streamlit==1.22.0
gradio==3.32.0

# Utilities
tqdm==4.65.0
matplotlib==3.7.0
seaborn==0.12.0
scikit-learn==1.2.0
python-dotenv==1.0.0
loguru==0.7.0

# Development
pytest==7.3.0
black==23.3.0
flake8==6.0.0 